Replace imageMaskOverlays module with unified imageMaskOverlay class; add presentation types
