+++ entrypoints/content/presentation/imagePredictions.ts
@@ -1,6 +1,6 @@
 import { markHandled, markProcessed } from '@/entrypoints/content/core/status';
 import { clearBlurBoxOverlay } from '@/entrypoints/content/presentation/boundingBox';
-import { clearImageMaskOverlay } from '@/entrypoints/content/presentation/imageMaskOverlays';
+import { imageMaskOverlay } from '@/entrypoints/content/presentation/imageMaskOverlay';
 import { removeInitialImageStyling } from '@/entrypoints/content/presentation/initialStyling';
 import { applyPredictionsStyling } from '@/entrypoints/content/presentation/predictionStyling';
 
@@ -52,7 +52,7 @@ export async function applyImagePredictionsToDom(
     // If this prediction has no detections, ensure overlays are cleared
     if (!pred.predictions || pred.predictions.length === 0) {
       for (const image of loadedImages) {
-        clearImageMaskOverlay(image);
+        imageMaskOverlay.clearMaskOverlay(image);
         clearBlurBoxOverlay(image);
         removeInitialImageStyling(image);
         markHandled(image, pred.src);
