+++ entrypoints/content/presentation/predictionStyling.ts
@@ -1,5 +1,5 @@
 import { createBlurBoxOverlays } from '@/entrypoints/content/presentation/boundingBox';
-import { createImageMaskOverlays } from '@/entrypoints/content/presentation/imageMaskOverlays';
+import { imageMaskOverlay } from '@/entrypoints/content/presentation/imageMaskOverlay';
 import { logger, extractUrlId } from '@/utils/logger';
 import { type IHostSettings, type IImagePrediction } from '@/utils/types';
 
@@ -27,15 +27,10 @@ export const applyPredictionsStyling = (
     if (imagePrediction && hostSettings.masking.blur) {
       const overlayPromise = new Promise<void>(resolve => {
         requestAnimationFrame(() => {
-          logger.withTag('predictionStyling').debug('Creating overlay', {
-            outline: hostSettings.outline,
-            imageSrc: extractUrlId(imageSrc),
-          });
-
           if (hostSettings.outline === 'bbox') {
             createBlurBoxOverlays(image, imagePrediction);
           } else if (hostSettings.outline === 'segment') {
-            createImageMaskOverlays(image, imagePrediction);
+            imageMaskOverlay.createMaskOverlay(image, imagePrediction);
           }
           resolve();
         });
