+++ utils/types/index.ts
@@ -3,3 +3,4 @@ export * from '@/utils/types/media';
 export * from '@/utils/types/model';
 export * from '@/utils/types/prediction';
 export * from '@/utils/types/messaging';
+export * from '@/utils/types/presentation';
